+++ lightkurve/io/tests/test_k2sff.py
@@ -4,6 +4,7 @@ from astropy.io import fits
 from numpy.testing import assert_array_equal
 
 from ..k2sff import read_k2sff_lightcurve
+from ... import search_lightcurve
 
 
 @pytest.mark.remote_data
@@ -18,3 +19,20 @@ def test_read_k2sff():
         # Are `time` and `flux` consistent with the FITS file?
         assert_array_equal(f[ext].data['T'], lc.time.value)
         assert_array_equal(f[ext].data['FCOR'], lc.flux.value)
+
+
+@pytest.mark.remote_data
+def test_search_k2sff():
+    """Can we search and download a K2SFF light curve?"""
+    # Try an early campaign
+    search = search_lightcurve("K2-18", author="K2SFF", campaign=1)
+    assert len(search) == 1
+    assert search.table["author"][0] == "K2SFF"
+    lc = search.download()
+    assert type(lc).__name__ == "KeplerLightCurve"
+    assert lc.campaign == 1
+    # Try a late campaign
+    lc = search_lightcurve("GJ 9827", author="K2SFF", campaign=19).download()
+    assert type(lc).__name__ == "KeplerLightCurve"
+    assert lc.targetid == 246389858
+    assert lc.campaign == 19